scripts/requirements.txt
google-cloud-bigquery>=3.0
bitarray>=2.0
